- Replace system IDPool by UUID
[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[ ] Bugfix
[ ] New Feature
[x] Cleanup
[x] Optimization
[ ] Documentation

+++ lib/client/event.h
@@ -203,7 +203,7 @@ namespace eq
         uint32_t type;
 
         /** The identifier of the entity emitting the event. */
-        uint32_t originator;
+        uint128_t originator;
 
         /** Data for the event corresponding to the event type. */
         union
